refactor: rename outrospector to outrospect
