+++ src/outrospect.ts
