+++ src/test.ts
@@ -1,5 +1,5 @@
 import { fetchIntrospection, saveJsonFormatted } from "./lib.ts";
-import { outrospect, outrospectionToJSON } from "./outrospector.ts";
+import { outrospect, outrospectionToJSON } from "./outrospect.ts";
 
 const introspection = await fetchIntrospection(
   "https://rickandmortyapi.com/graphql",
@@ -7,4 +7,4 @@ const introspection = await fetchIntrospection(
 // const {queryType, mutationType} = getQueryAndMutationTypes(introspection);
 const outrospection = outrospect(introspection);
 const outrospectionJSON = outrospectionToJSON(outrospection);
-saveJsonFormatted(outrospectionJSON, "outrospection.json");
+saveJsonFormatted(outrospectionJSON, "./out/outrospection.json");
